pyproject.toml
[project]
name = "pytuiplayer"
version = "0.1.0"
description = "A terminal-based music player built using Python, PyRadio, MPV, and Textual."
readme = "README.md"
authors = [
    { name = "ChetanKnowIt", email = "chetanamrao174@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-mpv>=1.0.8",
    "ruff>=0.14.10",
    "textual>=6.11.0",
]

[project.scripts]
pytuiplayer = "pytuiplayer:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
